# openapi.yaml
openapi: 3.0.3
info:
  title: Experience Loop (QR Edition) API
  description: |
    Experience Loop는 이벤트 참가자들의 실시간 피드백을 수집하고 
    AI 기반 인사이트를 생성하는 PWA 플랫폼입니다.
    
    주요 기능:
    - PDF 기반 이벤트 생성 및 AI 슬롯/질문 생성
    - 실시간 푸시 알림 시스템
    - 다양한 답변 입력 (텍스트/음성/사진)
    - AI 기반 답변 요약 및 태깅
    - 개인화된 인사이트 리포트 생성
  version: 1.0.0
  contact:
    name: Experience Loop Team
    email: support@experienceloop.com
servers:
  - url: http://localhost:3000
    description: 로컬 개발 환경
  - url: https://experience-loop.vercel.app
    description: 프로덕션 환경
paths:
...
  /api/reports/{eventId}/generate:
    post:
      summary: 참가자별 리포트 생성
      description: |
        이벤트의 모든 참가자에 대해 AI 기반 개인 인사이트 리포트를 생성합니다.
        OpenAI GPT-4를 사용하여 답변과 슬롯 데이터를 분석합니다.
      tags:
        - Reports
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 이벤트 ID
      responses:
        '200':
          description: 리포트 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "참가자별 리포트 생성이 완료되었습니다."
                  event:
                    $ref: '#/components/schemas/Event'
                  total_participants:
                    type: integer
                    description: 총 참가자 수
                    example: 25
                  reports_generated:
                    type: integer
                    description: 생성된 리포트 수
                    example: 25
                  reports:
                    type: array
                    items:
                      type: object
                      properties:
                        participant_id:
                          type: string
                          format: uuid
                        nickname:
                          type: string
                        team:
                          type: string
                        report_id:
                          type: string
                          format: uuid
                        webview_url:
                          type: string
                        pdf_url:
                          type: string
        '404':
          description: 이벤트를 찾을 수 없음
        '500':
          description: 서버 오류
  /api/reports/{participantId}:
    get:
      summary: 개인 리포트 조회
      description: 특정 참가자의 리포트를 조회합니다.
      tags:
        - Reports
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 참가자 ID
      responses:
        '200':
          description: 리포트 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: 리포트를 찾을 수 없음
...
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        password_hash:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        venue:
          type: string
        pdf_url:
          type: string
          format: uri
        status:
          type: string
          enum: [draft, published, completed]
        created_at:
          type: string
          format: date-time
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        content:
          type: object
          description: AI 생성 리포트 내용
          properties:
            timeline:
              type: array
              items:
                type: object
                properties:
                  slot_title:
                    type: string
                  snap:
                    type: string
                  photo:
                    type: string
                    nullable: true
            story:
              type: object
              properties:
                problem:
                  type: string
                solution:
                  type: string
                validation:
                  type: string
            mentoring:
              type: object
              properties:
                highlights:
                  type: array
                  items:
                    type: string
                applied:
                  type: array
                  items:
                    type: string
            strengths:
              type: array
              items:
                type: string
            improvements:
              type: array
              items:
                type: string
            next30:
              type: array
              items:
                type: object
                properties:
                  task:
                    type: string
                  due:
                    type: string
                    format: date
                  priority:
                    type: integer
                    minimum: 1
                    maximum: 5
            star:
              type: array
              items:
                type: object
                properties:
                  s:
                    type: string
                    description: Situation
                  t:
                    type: string
                    description: Task
                  a:
                    type: string
                    description: Action
                  r:
                    type: string
                    description: Result
            social:
              type: object
              properties:
                linkedin3:
                  type: string
                  description: 3줄 LinkedIn 요약
        webview_url:
          type: string
        pdf_url:
          type: string
        generated_at:
          type: string
          format: date-time
...